+++ lib/components/Accordion/types/index.ts
@@ -3,7 +3,7 @@ import React from "react";
 import { EIconName } from "../../Icons";
 import { ETooltipPosition } from "../../Tooltip";
 
-type TAccordionLevel = 1 | 2;
+export type TAccordionSize = "md" | "xl";
 
 export interface AccordionProps extends React.PropsWithChildren {
   name?: string;
@@ -18,7 +18,7 @@ export interface AccordionProps extends React.PropsWithChildren {
   classNameIcon?: string;
   classNameChildrenWrapper?: string;
   expandIconName?: EIconName;
-  level?: TAccordionLevel;
+  size?: TAccordionSize;
   infoTooltipText?: string;
   tooltipPosition?: ETooltipPosition;
   classNameBaseTooltipRoot?: string;